Revision after Shane and Pauls comments

+++ Calendar/BookingHelpers/ExcelWrite.ps1
@@ -1,4 +1,10 @@
-﻿. $PSScriptRoot\GraphInstaller.ps1
+﻿#========================
+#Excel needs Scriptlets helper functions to be loaded, in order to have access to graph and install Export-Excel module
+#========================
+
+
+. $PSScriptRoot\GraphInstaller.ps1
+. $PSScriptRoot\BookingGenericFunctions.ps1
 . $PSScriptRoot\BookingMBHelpers.ps1
 . $PSScriptRoot\BookingStaffHelpers.ps1
 . $PSScriptRoot\BookingServiceHelpers.ps1
@@ -10,7 +16,6 @@ function GetExcelParams($path, $tabName) {
     return @{
         Path                    = $path
         FreezeTopRow            = $true
-        #  BoldTopRow              = $true
         Verbose                 = $false
         TableStyle              = "Medium3"
         WorksheetName           = $tabName
@@ -28,62 +33,74 @@ function GetExcelParams($path, $tabName) {
 function ExcelWrite {
     param($identity)
 
+    Write-Host -ForegroundColor Gray "Checking Excel, Graph and bookings modules"
     CheckExcelModuleInstalled
     CheckGraphAuthModuleInstalled
     CheckGraphBookingsModuleInstalled
 
+    $modulesInstalled = CheckGraphModulesInstalled
+    Write-Verbose "Modules installed: $modulesInstalled"
+    if ($modulesInstalled -eq $false) {
+        Write-Host -ForegroundColor Red "Graph modules are not installed. Exiting..."
+        exit
+    }
+
     Connect-MgGraph -Scopes "User.Read.All", "Bookings.Read.All" -NoWelcome
 
-    $path = 'C:\Temp\BookingsPOC.xlsx'
+    Write-Host -ForegroundColor Green "Exporting to Excel..."
+
+    #Compose File name formatted as: BookingsSummary_<IdentityName>_<yyyy-MM-dd>.xlsx
+    $IdentityName = SplitIdentityFromEmail -email $identity
+    $DateStringToFileName = (Get-Date).ToString("yyyy-MM-dd")
+    $path = "BookingsSummary_$($IdentityName)_$($DateStringToFileName).xlsx"
     Remove-Item -Path $path -Force -EA Ignore
 
     $ExcelParamsArray = GetExcelParams -path $path -tabName "Business"
 
     Write-Host "Exporting Booking Business"
-    $a = Get-GraphBusiness -Identity $identity
-    $excel = $a | Export-Excel @ExcelParamsArray -PassThru
+    $BusinessData = Get-GraphBusiness -Identity $identity
+    $excel = $BusinessData | Export-Excel @ExcelParamsArray -PassThru
     Export-Excel -ExcelPackage $excel -WorksheetName "Business" -MoveToStart
 
     Write-Host "Exporting Bookings Page Settings"
-    $a = Get-GraphBusinessPage -Identity $identity
-    $a | Export-Excel -Path  $path -WorksheetName "Business" -StartRow 10 -Title "Page Settings" -TableStyle Medium3 -AutoSize
+    $BusinessPage = Get-GraphBusinessPage -Identity $identity
+    $BusinessPage | Export-Excel -Path  $path -WorksheetName "Business" -StartRow 10 -Title "Page Settings" -TableStyle Medium3 -AutoSize
 
 
     Write-Host "Exporting Booking Policy"
-    $a = Get-GraphBusinessBookingPolicy -Identity $identity
-    $a | Export-Excel -Path  $path -WorksheetName "Business" -BoldTopRow  -StartRow 10 -StartColumn 10 -Title "Scheduling Policy" -TableStyle Medium3 -AutoSize
+    $BookingPolicy = Get-GraphBusinessBookingPolicy -Identity $identity
+    $BookingPolicy | Export-Excel -Path  $path -WorksheetName "Business" -BoldTopRow  -StartRow 10 -StartColumn 10 -Title "Scheduling Policy" -TableStyle Medium3 -AutoSize
 
     Write-Host "Exporting Working Hours"
-    $a = Get-GraphBusinessWorkingHours -Identity $identity
-    $a | Export-Excel -Path  $path -WorksheetName "Business" -BoldTopRow  -StartRow 25  -Title "Working Hours" -TableStyle Medium24 -AutoSize
+    $WorkingHours = Get-GraphBusinessWorkingHours -Identity $identity
+    $WorkingHours | Export-Excel -Path  $path -WorksheetName "Business" -BoldTopRow  -StartRow 25  -Title "Working Hours" -TableStyle Medium24 -AutoSize
 
 
     Write-Host "Exporting Staff"
-    $staff = Get-GraphBookingsStaff -Identity $identity
-    $staff | Export-Excel -Path  $path -WorksheetName "Staff" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -Title "Staff List"
+    $Staff = Get-GraphBookingsStaff -Identity $identity
+    $Staff | Export-Excel -Path  $path -WorksheetName "Staff" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -Title "Staff List"
 
     Write-Host "Exporting Services"
-    $services = Get-GraphBookingsServices -Identity $identity
-    $services | Export-Excel -Path  $path -WorksheetName "Services" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -Title "Services"
+    $Services = Get-GraphBookingsServices -Identity $identity
+    $Services | Export-Excel -Path  $path -WorksheetName "Services" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -Title "Services"
 
 
     Write-Host "Exporting Custom Questions"
     #Get-MgBookingBusinessCustomQuestion -BookingBusinessId $identity | Export-Excel -Path  $path -WorksheetName "Custom Questions" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd
-    $customQuestions = Get-GraphBookingsCustomQuestions -Identity $identity
-    $customQuestions | Export-Excel -Path  $path -WorksheetName "Custom Questions" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -Title "Custom Questions"
+    $CustomQuestions = Get-GraphBookingsCustomQuestions -Identity $identity
+    $CustomQuestions | Export-Excel -Path  $path -WorksheetName "Custom Questions" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -Title "Custom Questions"
 
 
-    Write-Host "Exporting MessageTrace for the past 5 days"
+    Write-Host "Exporting MessageTrace for the past $($script:MessageTrackingDays) days"
     $script:MessageTrackingLogs | Export-Excel -Path  $path -WorksheetName "Message Traces" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -TableStyle Medium3
 
-    Write-Host "ExportingInternal data to Excel"
+    Write-Host "Exporting Membership Log data to Excel"
     $ArrayFromStaffLog = createArrayFromStaffLog -log $script:BookingStaffMembershipLog.MailboxLog
     $ArrayFromStaffLog | Export-Excel -Path  $path -WorksheetName "Membership Log" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize  -Title "Staff Membership Log"
 
-    Write-Host "ExportingInternal data to Excel"
+    Write-Host "Exporting Internal data to Excel"
     $script:TenantSettings   | Export-Excel -Path  $path -WorksheetName "Tenant Settings - Internal" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -TableStyle Medium3 #-Title "Tenant Settings - Internal"
     $script:bookingMBData  | Export-Excel -Path  $path -WorksheetName "Mailbox Data - Internal" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -TableStyle Medium3 #-Title "Mailbox Data - Internal"
-    #$script:StaffData  | Export-Excel -Path  $path -WorksheetName "Staff Data-Internal" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -TableStyle Medium3 #-Title "Staff Data - Internal"
     $script:MBPermissions | Export-Excel -Path  $path -WorksheetName "MB Permissions-Internal" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -TableStyle Medium3 #-Title "MB Permissions - Internal"
     $script:MBRecipientPermissions | Export-Excel -Path  $path -WorksheetName "Recipient Permissions-Internal" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize -TableStyle Medium3 #-tableTitle "Recipient Permissions - Internal"
 
@@ -104,25 +121,6 @@ function ExcelWrite {
     # }
     # Get-MgBookingBusinessStaffAvailability -BookingBusinessId $identity -BodyParameter $params | Export-Excel -Path  $path -WorksheetName "Availability 5 days" -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd
 
-
-
-    Write-Host "Exporting to Excel Completed"
+    Write-Host -ForegroundColor Green "Exporting to Excel Completed. File saved at $path"
 }
 
-
-function export-BookingStaffMembershipLog {
-    param(
-        [Parameter(Mandatory = $true)]
-        [string]$identity
-    )
-
-    $ArrayFromStaffLog = createArrayFromStaffLog -log $script:BookingStaffMembershipLog.MailboxLog
-
-    $MBStaffMembershipLogArray = @()
-    foreach ($log in $script:BookingStaffMembershipLog.MailboxLog.Split("`r`n")) {
-        $MBStaffMembershipLogArray += [PSCustomObject]@{
-            log = $log
-        }
-    }
-    $MBStaffMembershipLogArray | Export-Excel -Path  $path -WorksheetName "Membership Log Array" -TableStyle Medium3 -AutoFilter -FreezeTopRow -BoldTopRow -MoveToEnd -AutoSize  -Title "Staff Membership Log"
-}
